v1.3.5: - RNG updated to C++11 STL - Updated Project Configurations: changed exe to x64 default, and now x86 is named _x86 - More warnings quashed

+++ BridgeTrafficLoadSim/Lane.cpp
@@ -27,12 +27,12 @@ double CLane::GetNextArrivalTime() const
 	return m_NextArrivalTime;
 }
 
-int CLane::GetLaneID()
+size_t CLane::GetLaneID()
 {
 	return m_LaneIndex;
 }
 
-int CLane::GetDirection()
+size_t CLane::GetDirection()
 {
 	return m_Direction;
 }
